Lesson 24

+++ app/Http/Controllers/ThreadsController.php
@@ -104,12 +104,14 @@ class ThreadsController extends Controller
 
     /**
      * Remove the specified resource from storage.
-     *
+     * 
      * @param  \App\Thread  $thread
      * @return \Illuminate\Http\Response
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
         if( request()->wantsJson() ){
